Requests catch and ignore exeptions
This should never happend but we have experienced 
timeout exception once. This patch should take care of
that and next time we run the cronjob the module that
was skipped will be processed. That exception will not 
occure often. It s an anomaly that needs to be taken care of.