This will serve to load a yang-catalog without haveing to first time loading data to cache which too aproximately 8 minutes.
All that is moved to api-recovery container
and for the first time loading of yangcatalog.
Once we have cache loaded to redis - which is
a separate container we can run backend safely
and immidiately so if any problem will appear
and restart is needed it will restart immediatelly
and we don t have to wait 8 minutes anymore.

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>