Fix memory leak while parsing modules (#420)
Previously, the compilation status data could not
be garbage collected, at it was referenced in each
Modules object and each Modules object was
referenced from the Prepare object. We can let the
garbage collector know we will no longer need this
data after we are done parsing. This allows it to
safely free the memory.