Caching calls to the yangParser (#424)
* Fix memory leak while parsing modules

Previously, the compilation status data could not
be garbage collected, at it was referenced in each
Modules object and each Modules object was
referenced from the Prepare object. We can let the
garbage collector know we will no longer need this
data after we are done parsing. This allows it to
safely free the memory.

* Remove duplicate code

* Refactor populate.py

* Fix type hint

* Small typing related fixes

* Add type hints to receiver.py

* Rename member variables according to pep8

* More renaming

* Cache parsed modules

* Update tests

* Add debug strings to asserts

* Fix: remove extra arguments passed