fix: handle newer Git vers when checking for conflicting worktree
repoutil.Worktree is designed to clean up any existing worktree before
adding a new one. Currently, it relies on the error message produced
by Git to determine if there is one. However, the format of the error
message has recently changed [1], causing the test suite to fail.

This fix should work with both older and newer versions of Git.

1. https://github.com/git/git/commit/2a499264d39